Refactor json:api data shaper into a service
Cleans up the encoder and encapsulates this complicated logic into a
single place.
